add qod to index
