+++ public/index.html
@@ -7,7 +7,7 @@
 
     <!-- links to Bulma CSS Framework and CSS-->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.3/css/bulma.min.css">
-    <link rel="stylesheet" href="./assets/css/style.css">
+    <!-- <link rel="stylesheet" href="./assets/css/style.css"> -->
 
 </head>
 
@@ -69,5 +69,5 @@
 
     </section>
 
-    <script src="assets/js/script.js"></script>
+    <script src="js/index.js"></script>
 </body>
