--- public/index.js
@@ -1,27 +1,0 @@
-// Random Variables
-var questionNum = 0;
-var correctAns = 0;
-var questionIndex = 0;
-var scoreResult;
-// End of Random Variables
-
-// Connecting Elements
-var timer = document.getElementById("timer");
-var timeLeft = document.getElementById("timeLeft");
-var timesUp = document.getElementById("timesUp");
-
-var startDiv = document.getElementById("start");
-var startQuizBtn = document.getElementById("start-quiz-button");
-
-var questionDiv = document.getElementById("questionDiv");
-var questionTitle = document.getElementById("questionTitle");
-var choiceA = document.getElementById("btn0");
-var choiceB = document.getElementById("btn1");
-var choiceC = document.getElementById("btn2");
-var choiceD = document.getElementById("btn3");
-var answerCheck = document.getElementById("answerCheck");
-
-var summary = document.getElementById("summary");
-var submitInitialBtn = document.getElementById("submitInitialBtn");
-var initialInput = document.getElementById("initialInput");
-var everything = document.getElementById("everything");
